Start convert to native Go
