+++ .gitignore
@@ -1,4 +1,1 @@
-__pycache__/
-albius.so
-_build
 .buildconfig
