--- albius/lib/meson.build
@@ -1,3 +1,0 @@
-internal_lib = static_library('albius_internal',
-  'ffi_funcs.c',
-)
