--- meson.build
@@ -1,8 +1,0 @@
-project(
-  'Albius',
-  'c',
-  version: '0.1'
-)
-
-subdir('albius')
-subdir('python_bindings')
