--- python_bindings/lib/__init__.py
@@ -1,6 +1,0 @@
-import lib.ffi
-from lib.ffi import load_library
-
-
-albius = load_library()
-lib.ffi.setup_exports()
