--- python_bindings/lib/ffi/__init__.py
@@ -1,2 +1,0 @@
-from .utils import load_library
-from .exports import setup_exports
